Refactor gem file structure
