+++ lib/ruby-nessus/version1/host.rb
