+++ lib/ruby-nessus/version2/host.rb
