+++ spec/ruby-nessus/version_spec.rb
