+++ spec/spec_fixtures/xml.rb
